refactor: move getBearerToken method to Tokens class
